Add useRoleOptions hook
